LayoutTests/platform/chromium/fast/forms/calendar-picker/month-picker-mouse-operations-expected.txt

Tests if month picker mouse operations work as expected.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Check that page popup doesn't exist at first.
PASS document.getElementById("mock-page-popup") is null
PASS selectedMonth() is "2000-01"
PASS currentMonth() is "2000-01"
Check that hovering over an entry highlights it.
PASS selectedMonth() is "1999-12"
PASS currentMonth() is "2000-01"
PASS selectedMonth() is "2000-02"
PASS currentMonth() is "2000-01"
PASS selectedMonth() is "2000-02"
PASS currentMonth() is "2000-01"
Check that moving the mouse outside the popup de-highlights entries.
PASS selectedMonth() is null
Check that mouse click closes the popup and sets the value.
PASS document.getElementById("mock-page-popup") is null
PASS document.getElementById("month").value is "2000-01"
PASS successfullyParsed is true

TEST COMPLETE
